Add protobuf codegen decoder
